+++ noodles-bam/src/reader/query.rs
@@ -1,11 +1,12 @@
 use std::io::{self, Read, Seek};
 
 use noodles_bgzf as bgzf;
-use noodles_core::region::Interval;
+use noodles_core::{region::Interval, Position};
 use noodles_csi::{self as csi, binning_index::index::reference_sequence::bin::Chunk};
-use noodles_sam::{self as sam, alignment::RecordBuf};
+use noodles_sam::{self as sam, alignment::Record as _};
 
 use super::Reader;
+use crate::Record;
 
 /// An iterator over records of a BAM reader that intersects a given region.
 ///
@@ -18,7 +19,7 @@ where
     header: &'a sam::Header,
     reference_sequence_id: usize,
     interval: Interval,
-    record: RecordBuf,
+    record: Record,
 }
 
 impl<'a, R> Query<'a, R>
@@ -37,17 +38,15 @@ where
             header,
             reference_sequence_id,
             interval,
-            record: RecordBuf::default(),
+            record: Record::default(),
         }
     }
 
-    fn next_record(&mut self) -> io::Result<Option<RecordBuf>> {
-        self.reader
-            .read_record_buf(self.header, &mut self.record)
-            .map(|n| match n {
-                0 => None,
-                _ => Some(self.record.clone()),
-            })
+    fn next_record(&mut self) -> io::Result<Option<Record>> {
+        self.reader.read_record(&mut self.record).map(|n| match n {
+            0 => None,
+            _ => Some(self.record.clone()),
+        })
     }
 }
 
@@ -55,14 +54,21 @@ impl<'a, R> Iterator for Query<'a, R>
 where
     R: Read + Seek,
 {
-    type Item = io::Result<RecordBuf>;
+    type Item = io::Result<Record>;
 
     fn next(&mut self) -> Option<Self::Item> {
         loop {
             match self.next_record() {
                 Ok(Some(record)) => {
-                    if intersects(&record, self.reference_sequence_id, self.interval) {
-                        return Some(Ok(record));
+                    match intersects(
+                        &record,
+                        self.header,
+                        self.reference_sequence_id,
+                        self.interval,
+                    ) {
+                        Ok(true) => return Some(Ok(record)),
+                        Ok(false) => {}
+                        Err(e) => return Some(Err(e)),
                     }
                 }
                 Ok(None) => return None,
@@ -72,20 +78,27 @@ where
     }
 }
 
-pub(crate) fn intersects(
-    record: &RecordBuf,
+fn intersects(
+    record: &Record,
+    header: &sam::Header,
     reference_sequence_id: usize,
     region_interval: Interval,
-) -> bool {
+) -> io::Result<bool> {
     match (
-        record.reference_sequence_id(),
-        record.alignment_start(),
-        record.alignment_end(),
+        record
+            .reference_sequence_id()
+            .map(usize::try_from)
+            .transpose()?,
+        record
+            .alignment_start()
+            .map(Position::try_from)
+            .transpose()?,
+        record.alignment_end(header).transpose()?,
     ) {
         (Some(id), Some(start), Some(end)) => {
             let alignment_interval = (start..=end).into();
-            id == reference_sequence_id && region_interval.intersects(alignment_interval)
+            Ok(id == reference_sequence_id && region_interval.intersects(alignment_interval))
         }
-        _ => false,
+        _ => Ok(false),
     }
 }